requirements.txt
tensorflow
pillow
numpy
gradio
scikit-learn
